Refactor plugin organization
