+++ tests/endpoints/test_plugin_endpoint.py
@@ -2,9 +2,8 @@ import pytest
 import time
 from concurrent.futures import ThreadPoolExecutor
 from infrad.endpoints import plugin_endpoint as pe
-from infrad.plugin import Plugin
 import infrad.domain.models as m
-from infrad.shared.consts import JOB_STATE
+from infrad.shared.consts import JobState
 
 
 def test_plugin_endpoint_init():
@@ -19,7 +18,7 @@ def test_plugin_endpoint_sync_command_success(mocker):
     mock = mocker.Mock()
     mock.init_command = 'comm'
     mock.do_work.return_value = 'job completed successfully'
-    Plugin.plugins = [mock]
+    pe.Plugin.plugins = [mock]
 
 
     result = endpoint.exec('comm', ['arg1', 'arg2'], {'kwarg1': 'value1'})
@@ -27,7 +26,8 @@ def test_plugin_endpoint_sync_command_success(mocker):
     assert isinstance(result, m.CommandResult)
     assert result.message == 'Sync Execution Finished'
     assert result.data == "job completed successfully"
-    assert result.status == JOB_STATE.COMPLETED
+    assert result.status == JobState.COMPLETED
+
 
 def test_plugin_endpoint_sync_command_fail(mocker):
     endpoint = pe.PluginEndpoint()
@@ -35,7 +35,7 @@ def test_plugin_endpoint_sync_command_fail(mocker):
     mock = mocker.Mock()
     mock.init_command = 'comm'
     mock.do_work.return_value = 'job completed successfully'
-    Plugin.plugins = [mock]
+    pe.Plugin.plugins = [mock]
 
 
     result = endpoint.exec('missing_comm', ['arg1', 'arg2'], {'kwarg1': 'value1'})
@@ -43,7 +43,7 @@ def test_plugin_endpoint_sync_command_fail(mocker):
     assert isinstance(result, m.CommandResult)
     assert result.message == 'Command not found'
     assert result.data == None
-    assert result.status == JOB_STATE.FAILED
+    assert result.status == JobState.FAILED
 
 
 # @pytest.mark.skip
@@ -53,14 +53,15 @@ def test_plugin_endpoint_async_command_success(mocker):
     mock = mocker.Mock()
     mock.init_command = 'comm'
     mock.do_work.side_effect = lambda *args, **kwargs: time.sleep(0.01)
-    Plugin.plugins = [mock]
+    pe.Plugin.plugins = [mock]
 
     result = endpoint.exec('comm', ['arg1', 'arg2'], {'kwarg1': 'value1'})
 
     assert isinstance(result, m.CommandResult)
     assert result.message == 'Async Execution In Progress'
-    assert result.status == JOB_STATE.RUNNING
+    assert result.status == JobState.RUNNING
     assert result.data == None
+
 
 def test_plugin_endpoint_async_command_fail(mocker):
     endpoint = pe.PluginEndpoint()
@@ -68,12 +69,12 @@ def test_plugin_endpoint_async_command_fail(mocker):
     mock = mocker.Mock()
     mock.init_command = 'comm'
     mock.do_work.side_effect = ['out']
-    Plugin.plugins = [mock]
+    pe.Plugin.plugins = [mock]
 
 
     result = endpoint.exec('missing_comm', ['arg1', 'arg2'], {'kwarg1': 'value1'})
 
     assert isinstance(result, m.CommandResult)
     assert result.message == 'Command not found'
-    assert result.status == JOB_STATE.FAILED
+    assert result.status == JobState.FAILED
     assert result.data == None
