+++ tests/serializers/test_command_result_serializer.py
@@ -1,18 +1,19 @@
 import json
 from infrad.domain import models
+from infrad.shared.consts import JobState
 from infrad.serializers import command_result_serializer as crs
 
 
 def test_serialize_domain_command_result():
     command_result = models.CommandResult(
-        'status', 'message', data={
+        JobState.COMPLETED, 'message', data={
             'key1': 'value1',
             'key2': 'value2'
         })
 
     expected_json = """
         {
-            "status": "status",
+            "status": 0,
             "message": "message",
             "data": {
                 "key1": "value1",
